xsv-xct-001: reconcile package verification
